requirements.txt
fastapi>=0.115,<1.0
uvicorn>=0.30,<1.0
python-multipart>=0.0.9
httpx>=0.27,<1.0
python-dotenv>=1.0,<2.0
psycopg2-binary>=2.9,<3.0
Pillow>=10.0,<12.0
openai>=1.70,<2.0
